feat: create types folder and boxer folder inside components
