+++ src/components/ColumnBoxers.astro
@@ -1,5 +1,5 @@
 ---
-import type { Boxer } from "@/consts/boxers"
+import type { Boxer } from "@/types/Boxer"
 
 interface Props {
 	boxers: Boxer[]
