+++ src/components/boxer/DetailInfo.astro
