+++ src/consts/forecasts.ts
@@ -1,13 +1,4 @@
-export interface Forecast {
-	combatId: string
-	forecastData: ForecastData[]
-}
-
-export interface ForecastData {
-	boxerId: string
-	forecast: number
-	predictionsCount: number
-}
+import type { Forecast } from "@/types/Forecast"
 
 export const FORECASTS: Forecast[] = [
 	{
