+++ src/consts/sponsors.ts
@@ -1,35 +1,4 @@
-type SponsorId =
-	| "vicio"
-	| "revolut"
-	| "prime"
-	| "alsa"
-	| "spotify"
-	| "cerave"
-	| "grefusa"
-	| "el-pozo"
-	| "maxibon"
-	| "infojobs"
-type SponsorName =
-	| "Vicio"
-	| "Revolut"
-	| "Prime"
-	| "Alsa"
-	| "Spotify"
-	| "Cerave"
-	| "Grefusa"
-	| "ElPozo"
-	| "Maxibon"
-	| "InfoJobs"
-
-interface Sponsors {
-	id: SponsorId
-	name: SponsorName
-	url: string
-	image: {
-		width: number
-		height: number
-	}
-}
+import type { Sponsors } from "@/types/Sponsors"
 
 export const SPONSORS: Array<Sponsors> = [
 	{
