+++ src/pages/boxers/[id].astro
@@ -1,8 +1,8 @@
 ---
-import BoxerBigImage from "@/components/BoxerBigImage.astro"
-import BoxerClips from "@/components/BoxerClips.astro"
-import BoxerDetailInfo from "@/components/BoxerDetailInfo.astro"
-import BoxerSocialLink from "@/components/BoxerSocialLink.astro"
+import BoxerBigImage from "@/components/boxer/BigImage.astro"
+import BoxerClips from "@/components/boxer/Clips.astro"
+import BoxerDetailInfo from "@/components/boxer/DetailInfo.astro"
+import BoxerSocialLink from "@/components/boxer/SocialLink.astro"
 import SectionTitle from "@/components/SectionTitle.astro"
 
 import Instagram from "@/icons/instagram.astro"
@@ -12,7 +12,8 @@ import X from "@/icons/x.astro"
 import YouTube from "@/icons/youtube.astro"
 
 import { COUNTRIES } from "@/consts/countries"
-import { BOXERS, type Boxer } from "@/consts/boxers"
+import { BOXERS } from "@/consts/boxers"
+import type { Boxer } from "@/types/Boxer"
 import { COMBATS } from "@/consts/combats"
 import { FORECASTS } from "@/consts/forecasts"
 import Layout from "@/layouts/Layout.astro"
