+++ src/sections/Combat.astro
@@ -1,6 +1,6 @@
 ---
 import Typography from "@/components/Typography.astro"
-import type { Boxer } from "@/consts/boxers"
+import type { Boxer } from "@/types/Boxer"
 
 interface Props {
 	combatNumber: string
@@ -44,7 +44,6 @@ const { combatNumber, boxers } = Astro.props
 			class:list={[
 				"pointer-events-none flex flex-col items-center justify-center text-2xl font-semibold text-accent",
 			]}
-			tabindex="0"
 			><span class="text-primary opacity-80">Ver combate</span><span
 				class="-skew-x-6 text-xs font-light uppercase">Próximamente</span
 			></a
