+++ src/sections/Forecasts.astro
@@ -1,8 +1,8 @@
 ---
 import Action from "@/components/Action.astro"
-import ForecastBoxer from "@/components/ForecastBoxer.astro"
+import ForecastBoxer from "@/components/boxer/Forecast.astro"
 import Velocimetro from "@/components/Velocimetro.astro"
-import { type Boxer } from "@/consts/boxers"
+import type { Boxer } from "@/types/Boxer"
 
 export interface BoxerWithForecast extends Boxer {
 	forecast: number
