+++ src/sections/SelectYourBoxer.astro
@@ -1,7 +1,7 @@
 ---
 import ColumnBoxers from "@/components/ColumnBoxers.astro"
 import Typography from "@/components/Typography.astro"
-import BoxerBigImage from "@/components/BoxerBigImage.astro"
+import BoxerBigImage from "@/components/boxer/BigImage.astro"
 import { BOXERS } from "@/consts/boxers"
 import { COUNTRIES } from "@/consts/countries"
 
@@ -115,7 +115,7 @@ const boxerColumns = [
 
 <style>
 	.boxers-lists {
-		@apply relative md:h-[480px] xl:h-32 w-screen overflow-x-scroll;
+		@apply relative w-screen overflow-x-scroll md:h-[480px] xl:h-32;
 		scrollbar-width: none;
 	}
