+++ src/types/Boxer.ts
@@ -1,0 +1,26 @@
+export interface Boxer {
+	id: string
+	name: string
+	birthDate: Date
+	age: number
+	realName: string
+	weight: number
+	height: number
+	country: string
+	countryName?: string
+	versus: string | string[]
+	guard?: string
+	reach?: number
+	socials: {
+		twitch?: string
+		instagram?: string
+		twitter?: string
+		youtube?: string
+		tiktok?: string
+	}
+	clips?: Array<{
+		text: string
+		url: string
+	}>
+	rotate?: boolean
+}
