+++ src/types/Combat.ts
@@ -1,0 +1,4 @@
+export interface Combat {
+	id: string
+	boxers: string[] // es un string porque hace referencia a los ids de los boxeadores
+}
