+++ src/types/Forecast.ts
@@ -1,0 +1,10 @@
+export interface Forecast {
+	combatId: string
+	forecastData: ForecastData[]
+}
+
+export interface ForecastData {
+	boxerId: string
+	forecast: number
+	predictionsCount: number
+}
